+++ src/Services/DeliveryService/DeliveryService/Delivery.Core/Controllers/DeliveryController.cs
@@ -14,10 +14,10 @@ namespace DeliveryService.Delivery.Core.Controllers
 {
     [ApiController]
     [Route("api/v3/[controller]")]
-    public class DeliveryController(IDeliveryService            _deliveryService,                                      
-                                    IMapper                     _mapper,
-                                    ILogger<DeliveryController> _logger) : ControllerBase 
-    {   
+    public class DeliveryController(IDeliveryService _deliveryService,
+                                    IMapper _mapper,
+                                    ILogger<DeliveryController> _logger) : ControllerBase
+    {
         /// <summary>
         /// Получение доставки через Guid
         /// </summary>
@@ -37,11 +37,11 @@ namespace DeliveryService.Delivery.Core.Controllers
         /// <param name="request"></param>
         /// <returns></returns>
 
-        [HttpPost("/api/create-delivery")]             
+        [HttpPost("/api/create-delivery")]
         public async Task<ActionResult<CreateDeliveryResponse>> CreateAsync(CreateDeliveryRequest request, CancellationToken cancellationToken)
-        {            
+        {
             var delivery = await _deliveryService.CreateAsync(_mapper.Map<CreateDeliveryDto>(request), cancellationToken);
-            return Ok(_mapper.Map<CreateDeliveryResponse>(delivery));     
+            return Ok(_mapper.Map<CreateDeliveryResponse>(delivery));
         }
 
         /// <summary>
@@ -51,8 +51,8 @@ namespace DeliveryService.Delivery.Core.Controllers
         /// <returns></returns>
         [HttpPut("/api/update-delivery/{id}")]
         public async Task<ActionResult<EditDeliveryResponse>> UpdateAsync(Guid id, EditDeliveryRequest request, CancellationToken cancellationToken)
-        { 
-            var delivery = await _deliveryService.UpdateAsync(id, _mapper.Map<EditDeliveryRequest, UpdateDeliveryDto>(request), cancellationToken);            
+        {
+            var delivery = await _deliveryService.UpdateAsync(id, _mapper.Map<EditDeliveryRequest, UpdateDeliveryDto>(request), cancellationToken);
             return Ok(_mapper.Map<EditDeliveryResponse>(delivery));
         }
 
@@ -69,14 +69,14 @@ namespace DeliveryService.Delivery.Core.Controllers
         }
 
         /// <summary>
-        /// Получение статуса доставки через Guid
+        /// Получение статуса доставки через Guid заказа
         /// </summary>
         /// <param name="orderId"></param>
         /// <returns></returns>
-        [HttpGet("GetDeliveryStatusByOrderId")]
-        public async Task<IActionResult> GetDeliveryStatusByOrderId(Guid orderId, CancellationToken cancellationToken)
+        [HttpGet("/api/GetDeliveryStatus/{orderId}")]
+        public async Task<IActionResult> GetDeliveryStatus(Guid orderId, CancellationToken cancellationToken)
         {
-            var delivery = await _deliveryService.GetDeliveryByOrderIdAsync(orderId, cancellationToken);
+            var delivery = await _deliveryService.GetByIdAsync(orderId, cancellationToken);
             if (delivery != null)
                 return Ok(JsonSerializer.Serialize(delivery));
 
@@ -88,10 +88,13 @@ namespace DeliveryService.Delivery.Core.Controllers
         /// <param name="orderId"></param>
         /// <returns></returns>
         /// 
-        //[HttpGet("/api/getDeliveryStatusByOrderId/{}")]
-        //public async Task<IActionResult> GetDeliveryStatusByOrderId(Guid orderId)
+        //[HttpGet("/api/getDeliveryStatusByOrderId/{orderId}")]
+        //public async Task<IActionResult> GetDeliveryStatusByOrderId(Guid orderId, CancellationToken cancellationToken)
         //{
-        //    var delivery = await _deliveryRepository.GetDeliveryByOrderIdAsync(orderId);
-        //    if (
+        //    var delivery = await _deliveryService.GetDeliveryByOrderIdAsync(orderId, cancellationToken);
+        //    if (orderId != null)
+        //        return Ok(JsonSerializer.Serialize(delivery));
+        //}
     }
 }
+
