+++ src/Services/DeliveryService/DeliveryService/Delivery.Core/Mapping/DeliveryMappingProfile.cs
@@ -15,29 +15,17 @@ namespace DeliveryService.Delivery.Core.Mapping
         {
             CreateMap<CreateDeliveryDto, Domain.Entities.DeliveryEntities.Delivery>()
             .ForMember(dest => dest.Id, opt => opt.Ignore())
-            //.ForMember(dest => dest.DeliveryStatus, opt => opt.MapFrom(src => src.DeliveryStatus))
-            //.ForMember(dest => dest.OrderId, opt => opt.MapFrom(src => src.OrderId))
-            //.ForMember(dest => dest.OrderStatus, opt => opt.MapFrom(src => src.OrderStatus))                
-            //.ForMember(dest => dest.TotalQuantity, opt => opt.MapFrom(src => src.TotalQuantity))
-            //.ForMember(dest => dest.TotalPrice, opt => opt.MapFrom(src => src.TotalPrice))
-            //.ForMember(dest => dest.ShippingAddress, opt => opt.MapFrom(src => src.ShippingAddress))
-            //.ForMember(dest => dest.EstimatedDeliveryTime, opt => opt.MapFrom(src => src.EstimatedDeliveryTime))
-            //.ForMember(dest => dest.ActualDeliveryTime, opt => opt.MapFrom(src => src.ActualDeliveryTime));
-            .ForMember(dest => dest.DeliveryStatus, opt => opt.Ignore())
-            .ForMember(dest => dest.OrderId, opt => opt.Ignore())
-            .ForMember(dest => dest.OrderStatus, opt => opt.Ignore())            
-            .ForMember(dest => dest.TotalQuantity, opt => opt.Ignore())
-            .ForMember(dest => dest.TotalPrice, opt => opt.Ignore())
-            .ForMember(dest => dest.ShippingAddress, opt => opt.Ignore())
-            .ForMember(dest => dest.EstimatedDeliveryTime, opt => opt.Ignore())
-            .ForMember(dest => dest.ActualDeliveryTime, opt => opt.Ignore());
+            .ForMember(dest => dest.DeliveryStatus, opt => opt.MapFrom(src => src.DeliveryStatus))
+            .ForMember(dest => dest.OrderId, opt => opt.MapFrom(src => src.OrderId))
+            .ForMember(dest => dest.OrderStatus, opt => opt.MapFrom(src => src.OrderStatus))
+            .ForMember(dest => dest.TotalQuantity, opt => opt.MapFrom(src => src.TotalQuantity))
+            .ForMember(dest => dest.TotalPrice, opt => opt.MapFrom(src => src.TotalPrice))
+            .ForMember(dest => dest.ShippingAddress, opt => opt.MapFrom(src => src.ShippingAddress))
+            .ForMember(dest => dest.EstimatedDeliveryTime, opt => opt.MapFrom(src => src.EstimatedDeliveryTime))
+            .ForMember(dest => dest.ActualDeliveryTime, opt => opt.MapFrom(src => src.ActualDeliveryTime));
 
             CreateMap<Domain.Entities.DeliveryEntities.Delivery, CreateDeliveryResponse>();
-            CreateMap<CreateDeliveryRequest, CreateDeliveryDto>();
-            //CreateMap<DeliveryModel, DeliveryDto>();
-            //CreateMap<CreateDeliveryModel, CreateDeliveryDto>();
-            //CreateMap<DataAccess.Domain.Domain.Entities.Delivery, CreateDeliveryDto>();
-            //CreateMap<UpdateDeliveryModel, updateDeliveryDto>();
+            CreateMap<CreateDeliveryRequest, CreateDeliveryDto>();            
         }
     }
 }
