+++ src/Services/DeliveryService/DeliveryService/Kafka/Consumers/ConsumerBackgroundService.cs
@@ -1,7 +1,10 @@
 using System;
+using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
 using Confluent.Kafka;
+using DeliveryService.Delivery.BusinessLogic.Services.Delivery.Repositories;
+using DeliveryService.Domain.External.Entities;
 using DeliveryService.Kafka.Models;
 using DeliveryService.Kafka.Options;
 using Microsoft.Extensions.Hosting;
@@ -13,15 +16,32 @@ public abstract class ConsumerBackgroundService<TKey, TValue> : BackgroundServic
 {
     private ILogger _logger;
     private BaseConsumer<TKey, TValue> _baseConsumer;
+    private ApplicationOptions _applicationOptions;
+
     protected abstract string TopicName { get; }
+    private readonly IOrderRepository _orderRepository;
+    private readonly IDeliveryRepository _deliveryRepository;
 
     public ConsumerBackgroundService(
+        BaseConsumer<TKey, TValue> _baseConsumer,
         ILogger logger,
-        ApplicationOptions applicationOptions)
+        ApplicationOptions applicationOptions,
+        IOrderRepository orderRepository,
+        IDeliveryRepository deliveryRepository)
     {
+        _logger = logger;
+        _orderRepository = orderRepository;
+        _deliveryRepository = deliveryRepository;
+
         _baseConsumer = new BaseConsumer<TKey, TValue>(applicationOptions.KafkaOptions, logger, applicationOptions.GroupId);
         _logger = logger;
 
+    }
+
+    protected ConsumerBackgroundService(ILogger<NewOrderConsumer> logger, ApplicationOptions applicationOptions)
+    {
+        _logger = logger;
+        _applicationOptions = applicationOptions;
     }
 
     protected override async Task ExecuteAsync(CancellationToken stoppingToken)
@@ -31,6 +51,11 @@ public abstract class ConsumerBackgroundService<TKey, TValue> : BackgroundServic
         while (!stoppingToken.IsCancellationRequested)
         {
             await Consume(stoppingToken);
+            //var consumeResult = _baseConsumer.Consumer.Consume(stoppingToken);
+            //var order = JsonSerializer.Deserialize<Order>(consumeResult.Message.Value);
+            //Console.WriteLine("Order received in OrderConsumerService:\nData: {consumeResult.Message.Value}");
+            //// Update delivery status in the database.
+            //await UpdateOrderStatusAsync(order);
         }
 
         _baseConsumer.Consumer.Unsubscribe();
@@ -59,6 +84,18 @@ public abstract class ConsumerBackgroundService<TKey, TValue> : BackgroundServic
             var value = message is null ? message.Message.Value.ToString() : "No value";
             _logger.LogError(e, $"Error process message with key {key}, value {value}");
         }
+    }
+
+    private async Task UpdateOrderStatusAsync(Order order)
+    {
+        //await _orderRepository.UpdateStatusOrder(order);
+        //await _deliveryRepository.SaveDeliveryStatus(order);
+
+        //var orderObj = _orderContext.Orders.Find(order.Id);
+        //if (orderObj == null) return Task.CompletedTask;
+        //orderObj.Status = order.Status;
+        //_orderContext.Orders.Update(orderObj);
+        //return _orderContext.SaveChangesAsync();
     }
 
     protected abstract Task HandleAsync(ConsumeResult<TKey, TValue> message, CancellationToken cancellationToken);
